separated admin and member login UI
