+++ src/app/(public)/admin-login/page.tsx
@@ -1,0 +1,5 @@
+import AdminLoginForm from "@/components/forms/AdminLoginForm";
+
+export default function AdminLoginPage() {
+  return <AdminLoginForm />;
+}
