+++ src/app/(public)/member-login/page.tsx
@@ -1,0 +1,5 @@
+import MemberLoginForm from "@/components/forms/MemberLoginForm";
+
+export default function LoginPage() {
+  return <MemberLoginForm />;
+}
